ComponentModel: widen ByteMemoryPool total-bytes counter to long
_totalBytes was an int while the public TotalBytes is long; a pool of 64KB+
buffers can exceed int.MaxValue, overflowing the counter so TotalBytes reported
negative/garbage. Back it with a long.

+++ ComponentModel/ByteMemoryPool.cs
@@ -144,7 +144,9 @@ public class ByteMemoryPool : MemoryPool<byte>
 	/// </summary>
 	public int TotalCount => _totalCount;
 
-	private int _totalBytes;
+	// long-backed so the running total can exceed int.MaxValue without overflowing
+	// (the public TotalBytes is long; large pools of 64KB+ buffers can pass 2GB).
+	private long _totalBytes;
 
 	/// <summary>
 	/// Gets the total number of bytes currently stored in the pool across all sizes.
